Squish: Workaround bad focus issue when using Squish
When interacting with the file dialog QC seems to lose its
focus and never gets it completely back.
This in turn makes lots of short cuts useless and (context)
menu entries disappear or being disabled.
Avoid this by explicitly closing the popup of the line edit
if there is one.

Change-Id: I79f86b3133d2d0780d73d75e15416d59f35dfa1e
Reviewed-by: Robert Loehning <robert.loehning@qt.io>