Wizards: Fix JSON of qtquick2-extension wizard
"\." is an invalid backslash escape sequence in JSON. QJSON accepts
this though and silently "fixes" the string by removing the stray
'\' character. This is QTBUG-78443 now, as a conforming JSON parser
should reject invalid data.

The string as "fixed" by QJSON is also not what we wanted in the
first place.

Change-Id: I070999957c6a0c3d0f1753d907ac460e9e341fc1
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>