BareMetal: Fix auto detection of SDCC toolchain on Windows
The SDCC toolchain package can be provided as 32-bit
or as 64-bit installer. If the SDCC 64-bit package
will be installed on the 32-bit Windows, then it will
not be found in the system registry, because we use
the QSettings::NativeFormat. So, we need to check the
data for the 32-bit and 64-bit registry sequentially.

Change-Id: I1e7711bdde173eff21a7ba84f221d505a21709ca
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>