QmlDesigner: Fine tune import options dialog
Check boxes moved to before labels.
Groups separated by separator line instead of indentation.
If another non-boolean item has single boolean item as condition, and
the boolean item is not a condition to any other item, the items are
combined to single line. If this results in a single item group, the
group is omitted.
Grid utilizes available space better rather than be fixed size.

Change-Id: I70845c6352f26f3fc4c4bfa86e46a32e5b099b9b
Fixes: QDS-1117
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>