QmlDesigner: Fix eyeball in navigator for QQuick3DNode
The eyeball status for QQuickItem is handled in the form editor.
This does not work for QQuick3DNode.

* Adding support for 'invisible' in QmlVisualNode
  This allows to handle the 'invisible" state for
  both QQuickItem and QQuick3DNode.

* Adding support for 'invisible' in the NodeInstanceServer.
  This means that the nodes are actuallly made invisible,
  if the AuxiliaryData is set.

Even with this patch there is still a downside. We actually do
hide the QQuick3DNode, which is reflected by the value of 'visible'
in the property editor. This is not correct. The eybeball should not affect
the actual property. To properly implement this we need a way to hide an
item in the scene wihout changing the visibility property.

Task-number: QDS-1065
Change-Id: I713ec97bc89fdd7271b2174756f36ce3731cfd4e
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>