AbiWidget: Fix custom abi comboboxes' enabled state when mainComboBox is changed
When the mainComboBox is updated and its current selection changed via
setAbis, the custom ABI comboboxes weren't updated accordingly.

This is because in setAbis when doing d->m_abi->setCurrentIndex(), the
m_ignoreChanges guard is held which inhibit mainComboBoxChanged
function from doing anything for performance reasons. But the effect of
this is that custom ABI comboboxes were left in their previous state
(enabled or disabled).

This commit ensures the custom ABI comboboxes are updated according to
the new selected item in the mainComboBox at the end of setAbis
function.

Change-Id: Iedabde412a42985697d2cba33ddb6c154f10a68f
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>