AutoTest: Fix handling of parameterized boost tests
Setting the parameterized flag after the test case may
have been added to the found tests is useless.
Do it as early as possible to take it into account
when gathering information for location and type.

Beside this the filter for parameterized boost tests
had been wrong which in turn led to not executing
them at all.

Change-Id: I1a4345b2a751c79cc4fc6df8e201e9606c961aaf
Reviewed-by: David Schulz <david.schulz@qt.io>