BareMetal: Handle details preceding it message in the KEIL parser
We need to handle a multi-line error messages from the KEIL A51
assembler:

 http://www.keil.com/support/man/docs/a51/a51_er_nonfatalerror.htm

This assembler produces an error messages in the form like:

 00B0                 114         ljmp    usb_stub_isr     ; (B0) GPIF operation complete.
 *** _____________________________________^
 *** ERROR #A45 IN 114 (autovec_keil.a51, LINE 114): UNDEFINED SYMBOL
 00B4                 116         ljmp    usb_stub_isr     ; (B4) GPIF waveform.
 *** _____________________________________^
 *** ERROR #A45 IN 116 (autovec_keil.a51, LINE 116): UNDEFINED SYMBOL

where the details places before than a main message text.

This patch handles this error messages, but with an one limitation in that a details
should start with a four hexadecimal letters (in a common case it can not
contains that letters at all).

Change-Id: I0c9bf7ab72e1aef32498af5a1ef29c9221c185d5
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>