QmlDesigner: Keep qml2puppet alive, if local socket is not processed
In case Qt Creator did not process the events from local socket,
e.g. when a modal dialog was executed, we did kill the qml2puppet.

In this patch we first check if there are really no pending data
anymore. If there was pending data we restart the timers for each qml2puppet.

Change-Id: If19161636ee05835e01f082dd992e49f39d366bd
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>