ProjectExplorer: Do not return early on parsingFinished assert
This is a hack, not a proper solution. In theory, this here should
serve as a lock.

We currently get start - start - finish - finish sequences when
switching qmake targets quickly, keeping the run button disabled.

The whole setup here is unfortunate. Parsing inherently depend on
more than the project alone, it's roughly per-Target. So keeping
track of it in the Project complicates it.

Change-Id: I57d361ce21088d8e1ec53ce75f83ff3bd04851e1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>