RemoteLinux: Correct clean shutdown when debugging
If the inferior stops by itself and the debugger
tries to shutdown it had ended up waiting forever for
the additional workers (e.g. ports gatherer, channel
provider,...) to stop.
The debugger appeared as finished but the runworker was
still in a running state until the user would hit the stop
button of the application output pane.
Make the gdbserver essential to initiate a clean finish.

Change-Id: I16b2ebe5feadc88bb76ce34b49ac5d3456d0867c
Reviewed-by: hjk <hjk@qt.io>