Target setup page: Un-queue connection to kit manager
If we have a project with two enabled targets and we remove both of the
associated kits, one of the targets temporarily becomes the active one.
If the activeTargetChanged() signal gets queued, then the receiver will
be called with an invalid target pointer, as the target has been removed
in the meantime.
It is unclear why the signal was queued in the first place. Presumably
the original reasoning is no longer applicable after various
refactorings.
This amends (and effectively reverts) commit c13b20f160.

Task-number: QTCREATORBUG-23075
Change-Id: I3c63c763319734a4d52b3bb221919b20fd123dbe
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>