McuSupport Vendor specific flash/run CMake targets
The build system for QtMCU creates a CMake target
"flash_%1_and_bootloader" for STM targets but not for NXP targets.
Handle that (for now in a dirtry manner) in the run configuration.

In master branch we can add the API needed for the runconfiguration
to retrieve the list of available targets.

Change-Id: If2c9089a5dc004900424041898a6ff426ebb450b
Reviewed-by: Aurindam Jana <aurindam.jana@qt.io>