GenericProjectManager: Speed up adding lots of files
This operation was slowed down by a function that inserted a new element
into a sorted list without making use of the fact that the list is
already sorted.
As a test case, I added my whole ~/dev directory with ca 600,000 source
files to a generic project. With this patch, the time spent on building
up the new list went down from 90 minutes to one second.

Task-number: QTCREATORBUG-20652
Change-Id: If537e58a73cc5f09bb45d47f0beb2925048a2b14
Reviewed-by: hjk <hjk@qt.io>