BareMetal: Don't create default GDB server provider
... if a BareMetalDevice doesn't have any configured
provider yet.

Reason is that a 'Default GDB' provider does not means
that it is default. It just means, that this provider
can be attached to the remote running GDB server (it is
equivalent of "Attach to Running Debug Server").

Besides, in future, it is makes sense to remove this
provider at all as it is useless.

Change-Id: I2c5c59595dc7d724ac6e05177bbc6ae85cf67e2e
Reviewed-by: hjk <hjk@qt.io>