ClangTools: Fix converting old settings on Windows
As the settings keys are case-insensitive on Windows, the condition

  s->value(oldDiagnosticConfigIdKey).isNull()

in ClangToolsSettings::readSettings always evaluated to true, leaving
inconsistent settings around (new and old).

As a result, the QTC_ASSERT parallelRuns >= 1 was triggered in
ClangToolRunWorker::start() and nothing was analyzed.

Fix by using a truly different settings key after conversion.

Change-Id: I467f7d71298f240812a56dfcf7d6ca81728dee32
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>