ProjectExplorer: Fix FolderNode::addNestedNode
std::lower_bound returns the first entry that is greater than or
equal to the argument. It should never return anything smaller
than the argument, so the condition did not make sense.

This broke the project tree of Qt Creator's cmake project (and likely other
projects as well).

Invert the condition to make it clearer.

This amends commit 18ecbb9b14104e169021f0f245dd84728dceb4e0.

Change-Id: I99a002b5fb2c4ea6639b896ce8d2ef2b01f0815b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>