Output panes: Fix newline handling
On repeated invocations of OutputWindow::appendMessage(), embedded
newlines would simply disappear because of an off-by-one error.
The only reason this was not observed all the time is that the relevant
code is only run for process output, which usually comes in larger
chunks, so that the whole output would get procecessed in a single call
to appendMessage().

Fixes: QTCREATORBUG-23300
Change-Id: Ibd10d33462d1922671c3cd9325adbbf2841bfa06
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>