ResourceEditor: Fix priority for qrc sub-nodes
When adding a new file to a qrc file via the context menu, the wrong
project tree node was pre-selected in the wizard unless the user had run
the context menu either from the ".qrc" node or a qrc prefix node that
was *not* the one for the "/" prefix.
Now this works correctly for all nodes at or below the ".qrc" node.

Fixes: QTCREATORBUG-23210
Change-Id: Ia5e234e9861a480f973b76f8e3026ebc9c73ec35
Reviewed-by: hjk <hjk@qt.io>