Utils: Introduce InfoLabel
Utils::InfoLabel is meant to replace around 30 custom implementations of
a QHBoxLayout with an pixmap QLabel and a text QLabel.

Instead of fiddling with pixmaps, the user of InfoLabel can set an enum
for info/warning/etc... type. Internally, a globally shared QIcon is
used. The usage of QIcon ensures that @1x, @2x and hybrid cases are
handled.

Task-number: QTCREATORBUG-23346
Change-Id: I0f91a21e64f095db14837512263c4becbb2c13d7
Reviewed-by: Christian Stenger <christian.stenger@qt.io>