CheckableMessageBox: Forbid focus for text and pixmap
That allows changing between the buttons with the
cursor Left and Right keys, without loosing focus
from one of the buttons.

Before, that was impossible without intermediate
tab key pressing. E.g. when pressing Ctrl+W
to close the git commit editor, the focus was
outside the buttons. To select one of them,
Tab had to be pressed first, but after some
Key Right presses, the focus got away from the
buttons again and the game had to be repeated.

That behavior was different from QMessageBox,
where one of the buttons always has focus.

Change-Id: I9b12f794e8ba4aef1a68aaf249ba6dd6789cae4d
Reviewed-by: Eike Ziller <eike.ziller@qt.io>