FakeVim bugfix: Cmd-S shouldn't erase lines on Mac
On Mac, there's not only Ctrl but also Cmd. In FakeVim only checks for
Ctrl were being made when deciding whether to act on a keypress or not.
This was causing keypresses like <Cmd-S> to be registered as simply "S"
if it was not trapped by Qt Creator beforehand (e.g. if the "Save" menu
option was disabled).

Fixes: QTCREATORBUG-13392.
Change-Id: Ibd77c1e3c01f2cbc26ecd9a933ad5abbf96cbc34
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>